+++ src-tauri/src/indexer/processor.rs
@@ -2,7 +2,7 @@ use anyhow::{Context, Ok, Result};
 
 use crate::{database::{self, DbPool, files::{get_pending_files_for_space}}, vector_store::VectorStore};
 
-pub async fn process_space(pool: &DbPool, vector_store: &VectorStore, limit: i64) -> Result<()> {
+pub async fn process_space(pool: &DbPool, vector_store: &VectorStore, limit: i32) -> Result<()> {
     let pending_files = get_pending_files_for_space(pool, vector_store.space_id, limit)
         .await
         .context("failed to get pending indexed files")?;
